Implemented Image Upload and Implented Listing Creation Features, integrating Firebase Storage.
Summary:
- Added Listing edit and Listing creation page. The user can now create their own Listing, as well as input images for these Listings display.

Changes:
- Added buttons.dart, edit_listing.dart and listings_creation.dart.
- edit_listing holds the editing page implementation, while lisintgs_creation has the creation page implementation.
- Files which were changed:
 - MyListings.dart, added the Create Listing within the MyLisintgs Page.
 - house_tile.dart, added the edit button in the house_tile implementation. For now everyone can edit but later it will handle owner priveledges.
 - ListingService.dart, refactored the constructor to only allow one instance of the listingService, this way only one map of ID and Listing exists.
 -LodgingClass.dart & ProductClass.dart, changed price type to double, and added List attribute for image URLS.

Why:
- This feature will provide a more flexible approach to data linking front end with backend.

Related Issues:
- #133.
- #271.

+++ lib/Classes/ProductClass.dart
@@ -1,6 +1,6 @@
 class Product{
   String owner;
-  int price;
+  double price;
   String availability;
   String condition;
   String description;
@@ -19,7 +19,7 @@ class Product{
 
   //Getters
   String getOwner() => owner;
-  int getPrice() => price;
+  double getPrice() => price;
   String getAvailability() => availability;
   String getCondition() => condition;
   String getDescription() => description;
@@ -27,7 +27,7 @@ class Product{
 
   //Setters
   void setOwner(String owner) => this.owner = owner;
-  void setPrice(int price) => this.price = price;
+  void setPrice(double price) => this.price = price;
   void setAvailability(String availability) => this.availability = availability;
   void setCondition(String condition) => this.condition = condition;
   void setDescription(String description) => this.description = description;
